+++ apps/dash/1_display_only.py
@@ -15,4 +15,4 @@ app.layout = html.Div(
 )
 
 if __name__ == "__main__":
-    app.run(debug=True)
+    app.run(debug=True)  # pyright: ignore[reportUnknownMemberType]